NEWS added
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/fabia@59513 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/fabiac.c
@@ -1672,7 +1672,7 @@ SEXP spfabic(SEXP file_nameS, SEXP KS, SEXP alphaS, SEXP cycS, SEXP splS,SEXP sp
     }
 
 
-    nquant = floor(quant*n);
+    nquant = (int) floor(quant*n);
     spl = -spl;
     spz = -spz;
     in = 1.0/nn;
